requirements.txt
Django==2.2.1
djangorestframework==3.9.4
gunicorn==19.9.0
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0
